Revert "Revert "refactored""
This reverts commit 04edbaa3a79808753b311aba84394cadfdefc2be.

+++ src/main/java/com/gmail/spanteleyko/web/services/RoleService.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface RoleService {
-    public List<RoleDTO> get(int userId) throws SQLException;
+    public List<RoleDTO> get(Long userId) throws SQLException;
 
     public List<RoleDTO> get() throws SQLException;
